+++ R/exercise.R
@@ -34,10 +34,11 @@ setup_exercise_handler <- function(exercise_rx, session) {
   rv <- reactiveValues(triggered = 0, result = NULL)
 
   # debounce option to slow down successive exercise execution requests
-  debounce_ms <- getOption("tutorial.exercise.debounce", 0)   # value in milliseconds
-  if (debounce_ms > 0)
-    exercise_rx <- debounce(exercise_rx, debounce_ms)
-
+  debounce_s <- getOption("tutorial.exercise.debounce", 0) # value in seconds
+  if (is.numeric(debounce_s) && debounce_s > 0) {
+    # Convert debounce time to milliseconds
+    exercise_rx <- debounce(exercise_rx, debounce_s / 1000)
+  }
   # observe input
   observeEvent(exercise_rx(), {
     # get exercise from app